Move stack frames to separate files under lib/stack-frames
