+++ lib/dispatcher.js
@@ -18,13 +18,11 @@ const ValueCategory = Semantic.ValueCategory;
 const QueueItem = require('./default_dialog_queue');
 const Helpers = require('./helpers');
 
-const {
-    AcceptResult,
-    CancellationError,
-    QuestionStackFrame,
-    MultipleChoiceStackFrame,
-    ResultStackFrame
-} = require('./dialog_stack');
+const { CancellationError } = require('./errors');
+const { AcceptResult } = require('./stack-frames/base');
+const QuestionStackFrame = require('./stack-frames/simple-scalar-question');
+const MultipleChoiceStackFrame = require('./stack-frames/multiple-choice-question');
+const ResultStackFrame = require('./stack-frames/result');
 
 module.exports = class Dispatcher {
     constructor(manager, debug) {
